[1st commit] Sync your conf. between etcd & disks
Initially, I wanted to share some config files between load-balancer
pods (ingress controllers actually). Basically, TLS certs obtained
dynamically by my ingress controllers Traefik.

This little program runs easily as an init container (to pull the
config from etcd) or as a sidecar container (to keep the config in
sync). However, it can also be deployed (with Docker, or not) on
standard VMs, hardware...

For now, it simply supports etcd v3 but adding other key values stores
is just about implementing a small interface.

If I have time one day, I'd love to wrap all that in a Kubernetes
ThirdPartyResource in order to get a DynamicConfigMap resource.

IMPORTANT NOTE: this is intend to be used only for config files of a
reasonable size (a few MBs at the very most): the file contents
systematically are loaded in memory and KV stores aren't meant to
handle big files at all.